# packages/contracts/openapi/v1/openapi.api.json
{
  "components": {
    "schemas": {
...
      "Body_upload_avatar_api_account_avatar_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_avatar_api_account_avatar_post",
        "type": "object"
      },
...
      "CodeRequest": {
        "properties": {
          "code": {
            "maxLength": 16,
            "minLength": 4,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "CodeRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PasswordChangeRequest": {
        "properties": {
          "current_password": {
            "title": "Current Password",
            "type": "string"
          },
          "new_password": {
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "PasswordChangeRequest",
        "type": "object"
      },
      "PasswordRequest": {
        "properties": {
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "PasswordRequest",
        "type": "object"
      },
      "PreferencesUpdateRequest": {
        "description": "高级设置里需要服务端生效的用户偏好；纯本机偏好仍走 localStorage。",
        "properties": {
          "max_concurrent_runs": {
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Max Concurrent Runs",
            "type": "integer"
          }
        },
        "required": [
          "max_concurrent_runs"
        ],
        "title": "PreferencesUpdateRequest",
        "type": "object"
      },
      "ProfileUpdateRequest": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          }
        },
        "title": "ProfileUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenMathModel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/account/2fa/disable": {
      "post": {
        "operationId": "two_factor_disable_api_account_2fa_disable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Two Factor Disable",
        "tags": [
          "account"
        ]
      }
    },
    "/api/account/2fa/enable": {
      "post": {
        "operationId": "two_factor_enable_api_account_2fa_enable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Two Factor Enable",
        "tags": [
          "account"
        ]
      }
    },
    "/api/account/2fa/recovery-codes": {
      "post": {
        "operationId": "regenerate_recovery_codes_api_account_2fa_recovery_codes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate Recovery Codes",
        "tags": [
          "account"
        ]
      }
    },
...
    "/api/account/avatar": {
      "delete": {
        "description": "移除头像（幂等）：回落到姓名首字母。内容对象由存储层统一回收。",
        "operationId": "delete_avatar_api_account_avatar_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Delete Avatar",
        "tags": [
          "account"
        ]
      },
      "get": {
        "description": "返回当前登录用户的头像内容；仅本人可读，不按 user_id 暴露他人头像。",
        "operationId": "get_avatar_api_account_avatar_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Avatar",
        "tags": [
          "account"
        ]
      },
      "post": {
        "description": "上传头像：按魔数确认格式后内容寻址存储，库内只保留引用。",
        "operationId": "upload_avatar_api_account_avatar_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_avatar_api_account_avatar_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Avatar",
        "tags": [
          "account"
        ]
      }
    },
...
    "/api/account/password": {
      "post": {
        "operationId": "change_password_api_account_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Password",
        "tags": [
          "account"
        ]
      }
    },
    "/api/account/preferences": {
      "get": {
        "operationId": "get_preferences_api_account_preferences_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Preferences",
        "tags": [
          "account"
        ]
      },
      "put": {
        "description": "最大并发任务上限存服务端而不是浏览器：创建任务的闸门在这里校验，\n放 localStorage 改个缓存就能绕过。",
        "operationId": "update_preferences_api_account_preferences_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Preferences",
        "tags": [
          "account"
        ]
      }
    },
    "/api/account/profile": {
      "patch": {
        "operationId": "update_profile_api_account_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile",
        "tags": [
          "account"
        ]
      }
    },
...
    "/api/account/sessions/{session_id}": {
      "delete": {
        "operationId": "revoke_session_api_account_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Session",
        "tags": [
          "account"
        ]
      }
    },
...
  }
}